+++ microservice_account/src/AccountService.Application/Validators/LoginRequestValidator.cs
@@ -8,8 +8,7 @@ public sealed class LoginRequestValidator : AbstractValidator<LoginRequest>
     public LoginRequestValidator()
     {
         RuleFor(x => x.Email)
-            .NotEmpty()
-            .EmailAddress();
+            .NotEmpty();
 
         RuleFor(x => x.Password)
             .NotEmpty();